Make ircncasecmp count parameter a size_t, like strncmp.

+++ include/match.h
@@ -45,7 +45,7 @@ E int ToLower(int);
 E int ToUpper(int);
 
 E int irccasecmp(const char *, const char *);
-E int ircncasecmp(const char *, const char *, int);
+E int ircncasecmp(const char *, const char *, size_t);
 
 E void irccasecanon(char *);
 E void strcasecanon(char *);
